+++ test/toc-callback.js
@@ -6,7 +6,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback', t => {
   const callback = function (tocMarkdown, tocArray, tocHtml) {
     t.is(
       tocMarkdown,
-      '* [Heading](#heading)\n',
+      "* [Heading](#Heading)\n",
       'should work with disabled toc insertion + callback, returning toc only'
     );
     t.deepEqual(
@@ -14,7 +14,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback', t => {
       [
         {
           content: 'Heading',
-          anchor: 'heading',
+          anchor: "Heading",
           level: 1
         }
       ],
@@ -23,7 +23,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback', t => {
     t.is(
       tocHtml,
       `<ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>\n`,
       'should work with disabled toc insertion + callback, returning toc only'
     );
@@ -33,7 +33,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback', t => {
     mdIt('# Heading', {
       tocCallback: callback
     }),
-    '<h1 id="heading">Heading</h1>\n',
+    '<h1 id="Heading">Heading</h1>\n',
     'should work with disabled toc insertion + callback, returning heading only'
   );
 
@@ -44,7 +44,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback-toc', t => {
   const callback = function (tocMarkdown, tocArray, tocHtml) {
     t.is(
       tocMarkdown,
-      '* [Heading](#heading)\n',
+      "* [Heading](#Heading)\n",
       'should work with enabled toc insertion + callback, returning toc only'
     );
     t.deepEqual(
@@ -52,7 +52,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback-toc', t => {
       [
         {
           content: 'Heading',
-          anchor: 'heading',
+          anchor: "Heading",
           level: 1
         }
       ],
@@ -61,7 +61,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback-toc', t => {
     t.is(
       tocHtml,
       `<ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>\n`,
       'should work with enabled toc insertion + callback, returning toc only'
     );
@@ -77,10 +77,10 @@ test.cb('markdown-it-toc-and-anchor toc-callback-toc', t => {
       }
     ),
     `<p><ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>
 </p>
-<h1 id="heading">Heading</h1>\n`,
+<h1 id="Heading">Heading</h1>\n`,
     'should also work with callback, returning toc and heading'
   );
 
@@ -91,7 +91,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback-md', t => {
   const callback = function (tocMarkdown, tocArray, tocHtml) {
     t.is(
       tocMarkdown,
-      '* [Heading](#heading)\n',
+      "* [Heading](#Heading)\n",
       `should work with disabled toc insertion +
       callback in md options, returning toc only`
     );
@@ -100,7 +100,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback-md', t => {
       [
         {
           content: 'Heading',
-          anchor: 'heading',
+          anchor: "Heading",
           level: 1
         }
       ],
@@ -110,7 +110,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback-md', t => {
     t.is(
       tocHtml,
       `<ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>\n`,
       `should work with disabled toc insertion +
       callback in md options, returning toc only`
@@ -125,7 +125,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback-md', t => {
         tocCallback: callback
       }
     ),
-    '<h1 id="heading">Heading</h1>\n',
+    '<h1 id="Heading">Heading</h1>\n',
     `should work with disabled toc insertion +
     callback in md options, returning heading only`
   );
@@ -137,7 +137,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback-env', t => {
   const callback = function (tocMarkdown, tocArray, tocHtml) {
     t.is(
       tocMarkdown,
-      '* [Heading](#heading)\n',
+      "* [Heading](#Heading)\n",
       `should work with disabled toc insertion +
       callback in md render env, returning toc only`
     );
@@ -146,7 +146,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback-env', t => {
       [
         {
           content: 'Heading',
-          anchor: 'heading',
+          anchor: "Heading",
           level: 1
         }
       ],
@@ -156,7 +156,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback-env', t => {
     t.is(
       tocHtml,
       `<ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>\n`,
       `should work with disabled toc insertion +
       callback in md render env, returning toc only`
@@ -172,7 +172,7 @@ test.cb('markdown-it-toc-and-anchor toc-callback-env', t => {
         tocCallback: callback
       }
     ),
-    '<h1 id="heading">Heading</h1>\n',
+    '<h1 id="Heading">Heading</h1>\n',
     `should work with disabled toc insertion +
     callback in md render env, returning heading only`
   );