+++ test/toc.js
@@ -1,4 +1,5 @@
 import test from 'ava';
+import uslug from "uslug";
 
 import mdIt from './utils/md-it';
 
@@ -6,6 +7,8 @@ test('markdown-it-toc-and-anchor toc', t => {
   t.is(mdIt('', { toc: true }), '', 'should work with nothing');
 
   t.is(mdIt('@[toc]'), '<p></p>\n', 'should do nothing if not asked to');
+
+  t.is(mdIt("@[toc]", { toc: true }), "<p></p>\n", "should do nothing if not asked to");
 
   t.is(
     mdIt('@[toc]', { toc: true }),
@@ -24,11 +27,48 @@ test('markdown-it-toc-and-anchor toc', t => {
     ),
     `<p>test
 <ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>
 </p>
-<h1 id="heading">Heading</h1>\n`,
+<h1 id="Heading">Heading</h1>\n`,
     'should work with soft breaks'
+  );
+
+  t.is(
+    mdIt(
+      `@[toc]
+# Heading`,
+      {
+        toc: true,
+        appendIdToHeading: false
+      }
+    ),
+    `<p><ul class="markdownIt-TOC">
+<li><a href="#Heading">Heading</a></li>
+</ul>
+</p>
+<h1>Heading</h1>\n`,
+    "should not add id attribute if `appendIdToHeading` is falsy"
+  );
+
+
+  t.is(
+    mdIt(
+      `test
+[TOC]
+# Heading`,
+      {
+        toc: true,
+        tocPattern: /\[TOC\]/
+      }
+    ),
+    `<p>test
+<ul class="markdownIt-TOC">
+<li><a href="#Heading">Heading</a></li>
+</ul>
+</p>
+<h1 id="Heading">Heading</h1>\n`,
+    "should work with custom tocPattern"
   );
 
   t.is(
@@ -51,7 +91,7 @@ test('markdown-it-toc-and-anchor toc', t => {
       }
     ),
     `<p>@[tac]</p>
-<h1 id="heading">Heading</h1>\n`,
+<h1 id="Heading">Heading</h1>\n`,
     'should not generate toc with different tokens starting with @['
   );
 
@@ -65,11 +105,11 @@ test
       }
     ),
     `<p><ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>
 
 test</p>
-<h1 id="heading">Heading</h1>\n`,
+<h1 id="Heading">Heading</h1>\n`,
     `should work when there is a line break between @[toc]
 and next element in the same inline token`
   );
@@ -84,10 +124,10 @@ and next element in the same inline token`
       }
     ),
     `<p><ul class="test">
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>
 </p>
-<h1 id="heading">Heading</h1>\n`,
+<h1 id="Heading">Heading</h1>\n`,
     'should allow custom class'
   );
 
@@ -101,10 +141,10 @@ and next element in the same inline token`
       }
     ),
     `<p><ul>
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>
 </p>
-<h1 id="heading">Heading</h1>\n`,
+<h1 id="Heading">Heading</h1>\n`,
     /* eslint-disable max-len */
     'should handle not including default class in anchors when setting tocClassName to null'
   );
@@ -138,17 +178,17 @@ and next element in the same inline token`
       }
     ),
     `<p><ul class="markdownIt-TOC">
-<li><a href="#two">Two</a>
+<li><a href="#Two">Two</a>
 <ul>
-<li><a href="#three">Three</a></li>
+<li><a href="#Three">Three</a></li>
 </ul>
 </li>
 </ul>
 </p>
-<h1 id="heading">Heading</h1>
-<h2 id="two">Two</h2>
-<h3 id="three">Three</h3>
-<h1 id="one">One</h1>\n`,
+<h1 id="Heading">Heading</h1>
+<h2 id="Two">Two</h2>
+<h3 id="Three">Three</h3>
+<h1 id="One">One</h1>\n`,
     'should work when skipping first level'
   );
 
@@ -165,18 +205,18 @@ and next element in the same inline token`
       }
     ),
     `<p><ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a>
+<li><a href="#Heading">Heading</a>
 <ul>
-<li><a href="#two">Two</a></li>
+<li><a href="#Two">Two</a></li>
 </ul>
 </li>
-<li><a href="#one">One</a></li>
+<li><a href="#One">One</a></li>
 </ul>
 </p>
-<h1 id="heading">Heading</h1>
-<h2 id="two">Two</h2>
-<h3 id="three">Three</h3>
-<h1 id="one">One</h1>\n`,
+<h1 id="Heading">Heading</h1>
+<h2 id="Two">Two</h2>
+<h3 id="Three">Three</h3>
+<h1 id="One">One</h1>\n`,
     'should work when skipping last level'
   );
 
@@ -188,12 +228,12 @@ and next element in the same inline token`
       { toc: true }
     ),
     `<p><ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a></li>
-<li><a href="#heading-2">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
+<li><a href="#Heading-2">Heading</a></li>
 </ul>
 </p>
-<h1 id="heading">Heading</h1>
-<h1 id="heading-2">Heading</h1>\n`,
+<h1 id="Heading">Heading</h1>
+<h1 id="Heading-2">Heading</h1>\n`,
     'should work with smiliar levels and similar titles'
   );
 
@@ -207,12 +247,12 @@ and next element in the same inline token`
       { toc: true }
     ),
     `<p><ul class="markdownIt-TOC">
-<li><a href="#heading">'Heading' ?</a></li>
+<li><a href="#Heading">'Heading' ?</a></li>
 <li><a href="#lel">$.lel!</a></li>
 <li><a href="#lel-2">$.lel?</a></li>
 </ul>
 </p>
-<h1 id="heading">'Heading' ?</h1>
+<h1 id="Heading">'Heading' ?</h1>
 <h1 id="lel">$.lel!</h1>
 <h1 id="lel-2">$.lel?</h1>\n`,
     'should work with special chars'
@@ -281,7 +321,7 @@ and next element in the same inline token`
 ## SubHeading
 # Heading 2
 ### Deeper Heading`,
-      { toc: true }
+      { toc: true, slugify: (s) => uslug(s, { lower: true }) }
     ),
     `<p><ul class="markdownIt-TOC">
 <li><a href="#heading-1">Heading 1</a>
@@ -307,24 +347,24 @@ and next element in the same inline token`
   t.deepEqual(
     mdIt([ '# Heading', '# Heading', '# Heading' ], { resetIds: true }),
     [
-      '<h1 id="heading">Heading</h1>\n',
-      '<h1 id="heading">Heading</h1>\n',
-      '<h1 id="heading">Heading</h1>\n'
+      '<h1 id="Heading">Heading</h1>\n',
+      '<h1 id="Heading">Heading</h1>\n',
+      '<h1 id="Heading">Heading</h1>\n'
     ],
     `should return the same anchor hrefs for the same markdown headings with
-same names on different renderings with the same markdownIt instance when 
+same names on different renderings with the same markdownIt instance when
 resetIds is true`
   );
 
   t.deepEqual(
     mdIt([ '# Heading', '# Heading', '# Heading' ], { resetIds: false }),
     [
-      '<h1 id="heading">Heading</h1>\n',
-      '<h1 id="heading-2">Heading</h1>\n',
-      '<h1 id="heading-3">Heading</h1>\n'
+      '<h1 id="Heading">Heading</h1>\n',
+      '<h1 id="Heading-2">Heading</h1>\n',
+      '<h1 id="Heading-3">Heading</h1>\n'
     ],
-    `should return different anchor hrefs for the same markdown headings with 
-same names on different renderings with the same markdownIt instance when 
+    `should return different anchor hrefs for the same markdown headings with
+same names on different renderings with the same markdownIt instance when
 resetIds is false`
   );
 
@@ -345,23 +385,23 @@ resetIds is false`
     ),
     [
       `<p><ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>
 </p>
-<h1 id="heading">Heading</h1>\n`,
+<h1 id="Heading">Heading</h1>\n`,
       `<p><ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>
 </p>
-<h1 id="heading">Heading</h1>\n`,
+<h1 id="Heading">Heading</h1>\n`,
       `<p><ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>
 </p>
-<h1 id="heading">Heading</h1>\n`
+<h1 id="Heading">Heading</h1>\n`
     ],
-    `should return the same anchor hrefs for the same markdown headings with 
-same names on different renderings with the same markdownIt instance when 
+    `should return the same anchor hrefs for the same markdown headings with
+same names on different renderings with the same markdownIt instance when
 resetIds is true and toc is true`
   );
 
@@ -382,23 +422,24 @@ resetIds is true and toc is true`
     ),
     [
       `<p><ul class="markdownIt-TOC">
-<li><a href="#heading">Heading</a></li>
+<li><a href="#Heading">Heading</a></li>
 </ul>
 </p>
-<h1 id="heading">Heading</h1>\n`,
+<h1 id="Heading">Heading</h1>\n`,
       `<p><ul class="markdownIt-TOC">
-<li><a href="#heading-2">Heading</a></li>
+<li><a href="#Heading-2">Heading</a></li>
 </ul>
 </p>
-<h1 id="heading-2">Heading</h1>\n`,
+<h1 id="Heading-2">Heading</h1>\n`,
       `<p><ul class="markdownIt-TOC">
-<li><a href="#heading-3">Heading</a></li>
+<li><a href="#Heading-3">Heading</a></li>
 </ul>
 </p>
-<h1 id="heading-3">Heading</h1>\n`
+<h1 id="Heading-3">Heading</h1>\n`
     ],
-    `should return different anchor hrefs for the same markdown headings with 
-same names on different renderings with the same markdownIt instance when 
+    `should return different anchor hrefs for the same markdown headings with
+same names on different renderings with the same markdownIt instance when
 resetIds is false and toc is true`
   );
+
 });